Created the player entity, still needs to be implemented within graphics

+++ GecProject/Rectangle.h
@@ -7,10 +7,10 @@ public:
 	float m_xPos{ 0.f };
 	float m_yPos{ 0.f };
 
-	float m_height{ 0.f };
-	float m_width{ 0.f };
+	int m_height{ 0 };
+	int m_width{ 0 };
 
-	Rectangle(float xPos = 0.f, float yPos = 0.f, float height = 0.f, float width = 0.f) 
+	Rectangle(float xPos = 0.f, float yPos = 0.f, int height = 0, int width = 0) 
 		: m_xPos(xPos), m_yPos(yPos), m_height(height), m_width(width) {}
 
 	bool intersection(const Rectangle& other)
